requirements.txt
PyYAML==6.0.2
minio==7.2.15
openpyxl==3.1.5
